demo(react): remove drag event demo
