+++ packages/react-layout/README.md
@@ -3,7 +3,6 @@
 A react component that can easily use [egjs-infinitegrid's layouts](https://github.com/naver/egjs-infinitegrid)
 
 - [Drag Demo](https://naver.github.io/egjs-infinitegrid/assets/html/drag.html)
-- [Drag to indicator Demo](https://naver.github.io/egjs-infinitegrid/assets/html/indicator.html)
 
 
 ## Install
