+++ packages/react-layout/examples/drag.html
@@ -5,14 +5,22 @@
     </div>
 <div class="demobox">
     <h1>
-        React JustifiedLayout DragEvent
+        React JustifiedLayout Drag
     </h1>
     <div id="__react-content"></div>
 </div>
 
 <style>
+    body {
+        margin: 10px;
+        padding: 0;
+    }
     #__react-content {
         position: relative;
+    }
+    .wrapper {
+        position: relative;
+        width: 100%;
     }
     .item {
         position: absolute;
@@ -34,10 +42,40 @@
     .item .info {
         margin-top: 10px;
         font-weight: bold;
-        color: #777;
+        
     }
+    .item.is_drag {
+        transition: none;
+        opacity: 0.5;
+        z-index: 2;
+    }
+    .indicator.is_show {
+        display: block;
+    }
+    .indicator {
+        display: none;
+        position: absolute;
+        width: 10px;
+        background: #f55;
+        
+    }
+   .ghostImage {
+       position: absolute;
+       display: none;
+       opacity: 0.5;
+   }
+   .ghostImage img {
+       width: 100%;
+       height: 100%;
+   }
+   .ghostImage:before {
+       content: "";
+       position: absolute;
+       width: 100%;
+       height: 100%;
+       color: transparent;
+   }
 </style>
-
 <script src="./drag.js"></script>
 <script src="./dat.gui.js"></script>
 <script type="text/javascript">
