--- packages/react-layout/examples/indicator.html
@@ -1,105 +1,0 @@
-
-<meta name="viewport" content="width=device-width,initial-scale=1.0,minimum-scale=1.0,maximum-scale=1.0,user-scalable=no">
-<div class="github" style="position: fixed; width: 45px; height: 45px; z-index:10;top: 10px; right: 15px;">
-        <a href="https://github.com/naver/egjs-infinitegrid/tree/master/packages/react-layout" target="_blank"><img src="../../common/image/github.png" alt="github" style="width: 100%; height: 100%; border-radius: 50%; "></a>
-    </div>
-<div class="demobox">
-    <h1>
-        React JustifiedLayout Drag To Indicator(Over IE10)
-    </h1>
-    <div id="__react-content"></div>
-</div>
-
-<style>
-    body {
-        margin: 10px;
-        padding: 0;
-    }
-    #__react-content {
-        position: relative;
-    }
-    .wrapper {
-        position: relative;
-        width: 100%;
-    }
-    .item {
-        position: absolute;
-        width: 200px;
-        overflow: hidden;
-
-        transition: all ease 1s;
-    }
-    .item .thumbnail {
-        /* max-height: 300px; */
-        overflow: hidden;
-        /* border-radius: 8px; */
-        padding: 0;
-        border: 0;
-    }
-    .item .thumbnail img {
-        width: 100%;
-    }
-    .item .info {
-        margin-top: 10px;
-        font-weight: bold;
-        
-    }
-    .item.is_drag {
-        transition: none;
-        opacity: 0.5;
-        z-index: 2;
-    }
-    .indicator.is_show {
-        display: block;
-    }
-    .indicator {
-        display: none;
-        position: absolute;
-        width: 10px;
-        background: #f55;
-        
-    }
-   
-</style>
-<script src="./indicator.js"></script>
-<script src="./dat.gui.js"></script>
-<script type="text/javascript">
-    
-window.onload = function() {
-    var text = {
-        minColumn: 2,
-        maxColumn: 3,
-        margin: 10,
-        change: function () {
-            app.setState({
-                margin: text.margin,
-                column: [text.minColumn, text.maxColumn]
-            });    
-        }
-    }
-    var gui = new dat.GUI();
-    gui.width = 140;
-    gui.add(text, 'margin').onFinishChange(text.change);
-    gui.add(text, 'minColumn').onFinishChange(text.change);
-    gui.add(text, 'maxColumn').onFinishChange(text.change);
-};
-        
-    </script>
-
-<style>
-     @media (max-width: 600px) {
-        h1 {
-            font-size: 16px;
-            margin-right: 160px;
-        }
-    }
-    .dg.ac {
-        top: 60px;
-    }
-    .dg.ac .property-name {
-        width: 60%;
-    }
-    .dg.ac .c {
-        width: 40%;
-    }
-</style>
