+++ packages/react-layout/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@egjs/react-layout",
-  "version": "1.1.1",
+  "version": "1.2.0",
   "description": "A react component that can easily use egjs-infinitegrid's layouts.",
   "main": "index.js",
   "author": {
